Fix useSystemProperties
Previously the value set with Config.useSystemProperties(boolean) was
overriden before. This pull request removes the call to
HttpClientBuilder.useSystemProperties() from the ApacheClient
constructor, so it is only called when the appropriate config is set.
Also set the default value of useSystemProperties to true to not break
behavior from previous versions.